feat: step1+step2
